+++ javascript/packages/core/config/entities/targets/list.ts
@@ -1,19 +1,8 @@
 import { CellType } from '#core/components/cell/constants';
+import { INFERENCE_SERVER_STATE_CELL } from './shared';
 
 import type { ColumnConfig } from '#core/components/table/types/column-types';
 import type { ListViewConfig } from '#core/components/views/types';
-
-export const INFERENCE_SERVER_STATE = {
-  INVALID: 0,
-  INITIALIZED: 1,
-  CREATE_PENDING: 2,
-  SERVING: 3,
-  FAILED: 4,
-  DELETE_PENDING: 5,
-  CREATING: 6,
-  DELETING: 7,
-  DELETED: 8,
-} as const;
 
 const TARGET_COLUMNS: ColumnConfig<object>[] = [
   {
@@ -37,33 +26,7 @@ const TARGET_COLUMNS: ColumnConfig<object>[] = [
     type: CellType.TAG,
     accessor: () => 'Inference Server',
   },
-  {
-    id: 'status.state',
-    label: 'State',
-    type: CellType.STATE,
-    stateTextMap: {
-      [INFERENCE_SERVER_STATE.INVALID]: 'Invalid',
-      [INFERENCE_SERVER_STATE.INITIALIZED]: 'Initialized',
-      [INFERENCE_SERVER_STATE.CREATE_PENDING]: 'Create pending',
-      [INFERENCE_SERVER_STATE.SERVING]: 'Serving',
-      [INFERENCE_SERVER_STATE.FAILED]: 'Failed',
-      [INFERENCE_SERVER_STATE.DELETE_PENDING]: 'Delete pending',
-      [INFERENCE_SERVER_STATE.CREATING]: 'Creating',
-      [INFERENCE_SERVER_STATE.DELETING]: 'Deleting',
-      [INFERENCE_SERVER_STATE.DELETED]: 'Deleted',
-    },
-    stateColorMap: {
-      [INFERENCE_SERVER_STATE.INVALID]: 'gray',
-      [INFERENCE_SERVER_STATE.INITIALIZED]: 'blue',
-      [INFERENCE_SERVER_STATE.CREATE_PENDING]: 'blue',
-      [INFERENCE_SERVER_STATE.SERVING]: 'green',
-      [INFERENCE_SERVER_STATE.FAILED]: 'red',
-      [INFERENCE_SERVER_STATE.DELETE_PENDING]: 'blue',
-      [INFERENCE_SERVER_STATE.CREATING]: 'blue',
-      [INFERENCE_SERVER_STATE.DELETING]: 'blue',
-      [INFERENCE_SERVER_STATE.DELETED]: 'gray',
-    },
-  },
+  INFERENCE_SERVER_STATE_CELL,
 ];
 
 export const TARGET_LIST_CONFIG: ListViewConfig<object> = {
